Add mother class PetscFV
