+++ demos/multigrid/LaplacianFV.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "samurai_new/petsc_assembly.hpp"
+#include "samurai_new/petsc_FV.hpp"
 #include "samurai_new/gauss_legendre.hpp"
 #include "samurai_new/boundary.hpp"
 
@@ -12,68 +12,67 @@ constexpr T ce_pow(T num, unsigned int pow)
 }
 
 
+/**
+ * Set useful sizes to define the sparsity pattern of the matrix and perform the preallocation.
+ * @see samurai_new::petsc::nnz_per_row<cfg>(mesh), used in sparsity_pattern()
+*/
+template<std::size_t dim>
+using starStencilFV = samurai_new::petsc::PetscAssemblyConfig
+<
+    // ----  Stencil size 
+    // Cell-centered Finite Volume scheme:
+    // center + 1 neighbour in each Cartesian direction (2*dim directions) --> 1+2=3 in 1D
+    //                                                                         1+4=5 in 2D
+    1 + 2*dim,
+
+    // ----  Projection stencil size
+    // cell + 2^dim children --> 1+2=3 in 1D 
+    //                           1+4=5 in 2D
+    1 + (1 << dim), 
+
+    // ----  Prediction stencil size
+    // Here, order 1:
+    // cell + hypercube of 3 cells --> 1+3= 4 in 1D
+    //                                 1+9=10 in 2D
+    1 + ce_pow(3, dim), 
+
+    // ---- Index of the stencil center
+    // (as defined in FV_stencil())
+    1, 
+
+    // ---- Start index and size of contiguous cell indices
+    // (as defined in FV_stencil())
+    // Here, [left, center, right].
+    0, 3
+>;
 
 /**
  * @class LaplacianFV
  * Assemble the matrix for the problem -Lap(u)=f.
  * The matrix corresponds to the discretization of the operator -Lap by the Finite-Volume method.
 */
-template<class Field>
-class LaplacianFV : public samurai_new::petsc::PetscAssembly
+template<class Field, std::size_t dim=Field::dim, class cfg=starStencilFV<dim>>
+class LaplacianFV : public samurai_new::petsc::PetscFV<cfg, typename Field::mesh_t>
 {
 public:
     using field_t = Field;
     using Mesh = typename Field::mesh_t;
     using mesh_id_t = typename Mesh::mesh_id_t;
-    static constexpr std::size_t dim = Field::dim;
+    //static constexpr std::size_t dim = Field::dim;
 
-    Mesh& mesh;
-
-    LaplacianFV(Mesh& m) :
-        mesh(m)
+    LaplacianFV(Mesh& m) : 
+        samurai_new::petsc::PetscFV<cfg, Mesh>(m, FV_stencil(), FV_coefficients)
     {}
 
 private:
-    /**
-     * Set useful sizes to define the sparsity pattern of the matrix and perform the preallocation.
-     * @see samurai_new::petsc::nnz_per_row<cfg>(mesh), used in sparsity_pattern()
-    */
-    using cfg = samurai_new::petsc::PetscAssemblyConfig
-    <
-        // ----  Stencil size 
-        // Cell-centered Finite Volume scheme:
-        // center + 1 neighbour in each Cartesian direction (2*dim directions) --> 1+2=3 in 1D
-        //                                                                         1+4=5 in 2D
-        1 + 2*dim,
-
-        // ----  Projection stencil size
-        // cell + 2^dim children --> 1+2=3 in 1D 
-        //                           1+4=5 in 2D
-        1 + (1 << dim), 
-
-        // ----  Prediction stencil size
-        // Here, order 1:
-        // cell + hypercube of 3 cells --> 1+3= 4 in 1D
-        //                                 1+9=10 in 2D
-        1 + ce_pow(3, dim), 
-
-        // ---- Index of the stencil center
-        // (as defined in FV_stencil())
-        1, 
-
-        // ---- Start index and size of contiguous cell indices
-        // (as defined in FV_stencil())
-        // Here, [left, center, right].
-        0, 3
-    >;
 
     using Stencil = samurai_new::Stencil<cfg::scheme_stencil_size, dim>;
 
 
     /**
-     * @return the stencil of the Finite Volume scheme.
+     * @return the star stencil of the Finite Volume scheme.
     */
-    static inline Stencil FV_stencil()
+    static constexpr Stencil FV_stencil()
     {
         static_assert(dim >= 1 || dim <= 3, "Finite Volume stencil not implemented for this dimension");
 
@@ -111,49 +110,6 @@ private:
         return coeffs;
     }
 
-private:
-    PetscInt matrix_size() override
-    {
-        return static_cast<PetscInt>(mesh.nb_cells());
-    }
-
-    bool matrix_is_spd() override
-    {
-        // The projections/predictions kill the symmetry, so the matrix is spd only if the mesh is not refined.
-        return mesh.min_level() == mesh.max_level();
-    }
-
-    std::vector<PetscInt> sparsity_pattern() override
-    {
-        // Scheme + projection + prediction
-        return samurai_new::petsc::nnz_per_row<cfg>(mesh);
-    }
-
-    void assemble_scheme_on_uniform_grid(Mat& A) override
-    {
-        samurai_new::petsc::set_coefficients<cfg>(A, mesh, FV_stencil(), FV_coefficients);
-    }
-
-    void assemble_projection(Mat& A) override
-    {
-        static constexpr PetscInt number_of_children = (1 << dim);
-
-        samurai_new::for_each_cell_and_children<PetscInt>(mesh, 
-        [&] (PetscInt cell, const std::array<PetscInt, number_of_children>& children)
-        {
-            MatSetValue(A, cell, cell, 1, INSERT_VALUES);
-            for (unsigned int i=0; i<number_of_children; ++i)
-            {
-                MatSetValue(A, cell, children[i], -1./number_of_children, INSERT_VALUES);
-            }
-        });
-    }
-    
-    void assemble_prediction(Mat& A) override
-    {
-        assemble_prediction_impl(std::integral_constant<std::size_t, dim>{}, A, mesh);
-    }
-
 public:
     /**
      * @brief Creates a right-hand side in the form of a Field.
@@ -164,11 +120,11 @@ public:
     template<class Func>
     Field create_rhs(Func&& source_function, int source_poly_degree=-1)
     {
-        Field rhs("rhs", mesh);
+        Field rhs("rhs", this->mesh);
         rhs.fill(0);
         samurai_new::GaussLegendre gl(source_poly_degree);
 
-        samurai::for_each_cell(mesh, [&](const auto& cell)
+        samurai::for_each_cell(this->mesh, [&](const auto& cell)
         {
             const double& h = cell.length;
             rhs[cell] = gl.quadrature(cell, source_function) / pow(h, dim);
@@ -185,10 +141,10 @@ public:
     template<class Func>
     void enforce_dirichlet_bc(Field& field, Func&& dirichlet)
     {
-        if (&field.mesh() != &mesh)
+        if (&field.mesh() != &this->mesh)
             assert(false && "Not the same mesh");
 
-        samurai_new::foreach_cell_on_boundary(mesh, FV_stencil(), FV_coefficients,
+        samurai_new::foreach_cell_on_boundary(this->mesh, FV_stencil(), FV_coefficients,
         [&] (const auto& cell, const auto& towards_bdry, double out_coeff)
         {
             auto boundary_point = cell.face_center(towards_bdry);
@@ -207,127 +163,3 @@ public:
         return LaplacianFV(coarse_mesh);
     }
 };
-
-
-//-----------------------------//
-//     Assemble prediction     //
-//          (order 1)          //
-//-----------------------------//
-
-// 1D
-
-template<class Mesh>
-void assemble_prediction_impl(std::integral_constant<std::size_t, 1>, Mat& A, Mesh& mesh)
-{
-    using mesh_id_t = typename Mesh::mesh_id_t;
-
-    auto min_level = mesh[mesh_id_t::cells].min_level();
-    auto max_level = mesh[mesh_id_t::cells].max_level();
-    for(std::size_t level=min_level+1; level<=max_level; ++level)
-    {
-        auto set = samurai::intersection(mesh[mesh_id_t::cells_and_ghosts][level],
-                                        mesh[mesh_id_t::cells][level-1])
-                .on(level);
-
-        std::array<double, 3> pred{{1./8, 0, -1./8}};
-        set([&](const auto& i, const auto&)
-        {
-            for(int ii=i.start; ii<i.end; ++ii)
-            {
-                auto i_cell = static_cast<int>(mesh.get_index(level, ii));
-                MatSetValue(A, i_cell, i_cell, 1., INSERT_VALUES);
-
-                int sign_i = (ii & 1)? -1: 1;
-
-                for(int is = -1; is<2; ++is)
-                {
-                    auto i1 = static_cast<int>(mesh.get_index(level - 1, (ii>>1) + is));
-                    double v = -sign_i*pred[is + 1];
-                    MatSetValue(A, i_cell, i1, v, INSERT_VALUES);
-                }
-
-                auto i0 = static_cast<int>(mesh.get_index(level - 1, (ii>>1)));
-                MatSetValue(A, i_cell, i0, -1., INSERT_VALUES);
-            }
-        });
-    }
-}
-
-
-// 2D
-
-template<class Mesh>
-void assemble_prediction_impl(std::integral_constant<std::size_t, 2>, Mat& A, Mesh& mesh)
-{
-    using mesh_id_t = typename Mesh::mesh_id_t;
-
-    auto min_level = mesh[mesh_id_t::cells].min_level();
-    auto max_level = mesh[mesh_id_t::cells].max_level();
-    for(std::size_t level=min_level+1; level<=max_level; ++level)
-    {
-        auto set = samurai::intersection(mesh[mesh_id_t::cells_and_ghosts][level],
-                                         mesh[mesh_id_t::cells][level-1])
-                .on(level);
-
-        std::array<double, 3> pred{{1./8, 0, -1./8}};
-        set([&](const auto& i, const auto& index)
-        {
-            auto j = index[0];
-            int sign_j = (j & 1)? -1: 1;
-
-            for(int ii=i.start; ii<i.end; ++ii)
-            {
-                auto i_cell = static_cast<PetscInt>(mesh.get_index(level, ii, j));
-                MatSetValue(A, i_cell, i_cell, 1, INSERT_VALUES);
-
-                int sign_i = (ii & 1)? -1: 1;
-
-                for(int is = -1; is<2; ++is)
-                {
-                    auto i1 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1), (j>>1) + is));
-                    MatSetValue(A, i_cell, i1, -sign_j*pred[is + 1], INSERT_VALUES);
-
-                    i1 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1) + is, (j>>1)));
-                    MatSetValue(A, i_cell, i1, -sign_i*pred[is + 1], INSERT_VALUES);
-                }
-
-                auto i1 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1) - 1, (j>>1) - 1));
-                auto i2 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1) + 1, (j>>1) - 1));
-                auto i3 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1) - 1, (j>>1) + 1));
-                auto i4 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1) + 1, (j>>1) + 1));
-
-                MatSetValue(A, i_cell, i1, sign_i*sign_j*pred[0]*pred[0], INSERT_VALUES);
-                MatSetValue(A, i_cell, i2, sign_i*sign_j*pred[2]*pred[0], INSERT_VALUES);
-                MatSetValue(A, i_cell, i3, sign_i*sign_j*pred[0]*pred[2], INSERT_VALUES);
-                MatSetValue(A, i_cell, i4, sign_i*sign_j*pred[2]*pred[2], INSERT_VALUES);
-
-                auto i0 = static_cast<PetscInt>(mesh.get_index(level - 1, (ii>>1), (j>>1)));
-                MatSetValue(A, i_cell, i0, -1, INSERT_VALUES);
-            }
-        });
-    }
-}
-
-
-// 3D
-
-template<class Mesh>
-void assemble_prediction_impl(std::integral_constant<std::size_t, 3>, Mat& /*A*/, Mesh& mesh)
-{
-    using mesh_id_t = typename Mesh::mesh_id_t;
-
-    auto min_level = mesh[mesh_id_t::cells].min_level();
-    auto max_level = mesh[mesh_id_t::cells].max_level();
-    for(std::size_t level=min_level+1; level<=max_level; ++level)
-    {
-        auto set = samurai::intersection(mesh[mesh_id_t::cells_and_ghosts][level],
-                                         mesh[mesh_id_t::cells][level-1])
-                .on(level);
-
-        //std::array<double, 3> pred{{1./8, 0, -1./8}};
-        set([&](const auto&, const auto&)
-        {
-            assert(false && "non implemented");
-        });
-    }
-}
